fix(test): fix narrowing conversion warning in GCC build
Explicitly cast size_t subtraction result to uint32_t to suppress
-Werror=conversion on GCC.

+++ tests/unit/actor/test_circuit_breaker_result.cpp
@@ -113,8 +113,8 @@ class CircuitBreakerResultTest : public ::testing::Test {
         }
         // If we need more than kNumBuckets, stack them in bucket 0
         if (needed > FailureRateTracker::kNumBuckets) {
-            ft_->failure_buckets[0] =
-                needed - (FailureRateTracker::kNumBuckets - 1);
+            ft_->failure_buckets[0] = static_cast<uint32_t>(
+                needed - (FailureRateTracker::kNumBuckets - 1));
         }
         ft_->last_bucket_advance = std::chrono::steady_clock::now();
         ft_->bucket_interval_ms = window_ms / FailureRateTracker::kNumBuckets;
